requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-ollama===0.2.2
langchain-core==0.3.41
langchain-text-splitters==0.3.6
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
psycopg2==2.9.10
GitPython==3.1.44
faiss-cpu==1.10.0
pgvector==0.3.6
alembic==1.15.1